for droid tmrl

+++ src/dsrl_openpi/models/pi0.py
@@ -66,7 +66,8 @@ def posemb_sincos(
 
 @dataclasses.dataclass(frozen=True)
 class Pi0Config(_model.BaseModelConfig):
-    dtype: str = "bfloat16"
+    # dtype: str = "bfloat16"
+    dtype: str = "float32"
     paligemma_variant: _gemma.Variant = "gemma_2b"
     action_expert_variant: _gemma.Variant = "gemma_300m"
